+++ src/main/java/conspire/cards/curse/Blindness.java
@@ -1,0 +1,57 @@
+package conspire.cards.curse;
+
+import com.megacrit.cardcrawl.actions.common.ApplyPowerAction;
+import com.megacrit.cardcrawl.actions.common.SetDontTriggerAction;
+import com.megacrit.cardcrawl.cards.AbstractCard;
+import com.megacrit.cardcrawl.cards.CardQueueItem;
+import com.megacrit.cardcrawl.characters.AbstractPlayer;
+import com.megacrit.cardcrawl.core.CardCrawlGame;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
+import com.megacrit.cardcrawl.localization.CardStrings;
+import com.megacrit.cardcrawl.monsters.AbstractMonster;
+import com.megacrit.cardcrawl.relics.BlueCandle;
+
+import basemod.abstracts.CustomCard;
+import conspire.Conspire;
+import conspire.powers.BlindnessPower;
+
+public class Blindness extends CustomCard {
+    public static final String ID = "conspire:Blindness";
+    private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
+    public static final String NAME = cardStrings.NAME;
+    public static final String DESCRIPTION = cardStrings.DESCRIPTION;
+    private static final int COST = -2;
+
+    public Blindness() {
+        super(ID, NAME, Conspire.cardImage(ID), COST, DESCRIPTION, CardType.CURSE, CardColor.CURSE, CardRarity.CURSE, CardTarget.NONE);
+    }
+
+    @Override
+    public void use(AbstractPlayer p, AbstractMonster m) {
+        if (!this.dontTriggerOnUseCard && p.hasRelic(BlueCandle.ID)) {
+            this.useBlueCandle(p);
+        } else {
+            AbstractDungeon.actionManager.addToTop(new ApplyPowerAction(AbstractDungeon.player, AbstractDungeon.player, new BlindnessPower((AbstractDungeon.player))));
+        }
+    }
+
+    @Override
+    public void triggerWhenDrawn() {
+        AbstractDungeon.actionManager.addToBottom(new SetDontTriggerAction(this, false));
+    }
+
+    @Override
+    public void triggerOnEndOfTurnForPlayingCard() {
+        this.dontTriggerOnUseCard = true;
+        AbstractDungeon.actionManager.cardQueue.add(new CardQueueItem(this, true));
+    }
+
+    @Override
+    public AbstractCard makeCopy() {
+        return new Blindness();
+    }
+
+    @Override
+    public void upgrade() {
+    }
+}