+++ src/main/java/conspire/powers/BlindnessPower.java
@@ -1,0 +1,23 @@
+package conspire.powers;
+
+import com.megacrit.cardcrawl.core.AbstractCreature;
+import com.megacrit.cardcrawl.core.CardCrawlGame;
+import com.megacrit.cardcrawl.localization.PowerStrings;
+
+public class BlindnessPower extends AbstractConspirePower {
+    public static final String POWER_ID = "conspire:Blindness";
+    private static final PowerStrings powerStrings = CardCrawlGame.languagePack.getPowerStrings(POWER_ID);
+    public static final String NAME = powerStrings.NAME;
+    public static final String[] DESCRIPTIONS = powerStrings.DESCRIPTIONS;
+
+    public BlindnessPower(AbstractCreature owner) {
+        super(POWER_ID, NAME, owner);
+        this.type = PowerType.DEBUFF;
+        this.updateDescription();
+    }
+
+    @Override
+    public void updateDescription() {
+        this.description = DESCRIPTIONS[0];
+    }
+}