+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### v1.2.0
 * Mod renamed to Conspire, to become more general than just monsters
-* Added cards: 2 red, 2 green, 1 blue, 1 colorless, 3 special
+* Added cards: 2 red, 2 green, 1 blue, 1 colorless, 3 special, 1 curse
 * Added potions: 2 potions
 * Added relics: 1 rare, 3 boss
 * Added events: Lone Ghost, Investor, (changed) Duplicator shrine